// packages/sn-table/src/types.ts
export type Align = "left" | "center" | "right";

export interface ViewService {
  rowPartialHeight?: number;
  scrollLeft?: number;
  scrollTopRatio?: number;
  qTop: number;
  qHeight: number;
  visibleLeft?: number;
  visibleWidth?: number;
  visibleTop?: number;
  visibleHeight?: number;
  rowsPerPage?: number;
  page?: number;
  viewState?: ViewState;
  estimatedRowHeight: number;
}
// ...
export interface Cell {
  qText?: string;
  qNum?: number | string;
  align: Align;
  qAttrExps?: EngineAPI.INxAttributeExpressionValues;
  qElemNumber: number;
  rowIdx: number;
  colIdx: number;
  pageRowIdx: number;
  pageColIdx: number;
  selectionColIdx: number;
  isSelectable: boolean;
  isLastRow: boolean;
  isLastColumn?: boolean;
  isNumeric?: boolean;
}

export interface Row {
  // for the row key, string is needed
  [key: string]: Cell | string;
}

export type SortDirection = "A" | "D";

export interface Column {
  id: string;
  isDim: boolean;
  qLibraryId?: string;
  fieldId: string;
  isLocked: boolean;
  colIdx: number;
  pageColIdx: number;
  selectionColIdx: number;
  isActivelySorted: boolean;
  label: string;
  headTextAlign: Align;
  totalsTextAlign: Align;
  bodyTextAlign: "auto" | Align;
  stylingIDs: string[];
  sortDirection: SortDirection;
  qReverseSort: boolean;
  totalInfo: string;
  qApprMaxGlyphCount: number;
  columnWidth: ColumnWidth;
}

export type TotalsPosition = { atTop: boolean; atBottom: boolean };

export type TableData = {
  totalColumnCount: number;
  totalRowCount: number;
  totalPages: number;
  paginationNeeded: boolean;
  rows: Row[];
  columns: Column[];
  totalsPosition: TotalsPosition;
};

export interface PageInfo {
  page: number;
  rowsPerPage: number;
  rowsPerPageOptions: number[];
}

export type SetPageInfo = stardust.SetStateFn<PageInfo>;

export interface BodyColors {
  borderColor: string;
}

export interface ExtendedSelectionAPI extends stardust.ObjectSelections {
  on(eventType: string, callback: () => void): void;
  removeListener(eventType: string, callback: () => void): void;
}

export interface ViewState {
  rowPartialHeight: number;
  scrollLeft: number;
  scrollTopRatio?: number;
  visibleTop: number;
  visibleHeight: number;
  visibleLeft?: number;
  visibleWidth?: number;
  rowsPerPage?: number;
  page?: number;
  isMultiPage?: boolean;
  maxLineCount?: number;
  usePagination?: boolean;
  skipTotals?: boolean;
  skipHeader?: boolean;
}